# openapi.yaml
openapi: 3.0.3
info:
  title: JWT Authentication API
  description: Authentication service with JWT tokens using RSA-256 signing
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com

servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT tokens
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=<token>; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=604800
              description: Refresh token set as httpOnly cookie
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token from cookie
      operationId: refreshToken
      security: []
      parameters:
        - in: cookie
          name: refresh_token
          schema:
            type: string
          required: true
          description: Refresh token from httpOnly cookie
      responses:
        '200':
          description: Token successfully refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=<new_token>; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=604800
              description: New refresh token set as httpOnly cookie
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Logout current user and blacklist tokens
      operationId: logoutUser
      parameters:
        - in: cookie
          name: refresh_token
          schema:
            type: string
          required: false
          description: Refresh token to blacklist
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: Successfully logged out
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=0
              description: Clear refresh token cookie
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/verify:
    get:
      tags:
        - Authentication
      summary: Verify token validity
      description: Verify if the current access token is valid
      operationId: verifyToken
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/revoke:
    post:
      tags:
        - Authentication
      summary: Revoke specific token
      description: Admin endpoint to revoke a specific token
      operationId: revokeToken
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeRequest'
      responses:
        '200':
          description: Token successfully revoked
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: Token successfully revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in Authorization header

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: user@example.com
          description: User email address
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
          example: SecureP@ss123
          description: Password must contain uppercase, lowercase, number, and special character
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_-]+$'
          example: john_doe
          description: Username for display

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
          description: User email address
        password:
          type: string
          format: password
          example: SecureP@ss123
          description: User password

    RevokeRequest:
      type: object
      required:
        - token_id
      properties:
        token_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique identifier of the token to revoke
        reason:
          type: string
          maxLength: 255
          example: Security breach detected
          description: Optional reason for revocation

    UserResponse:
      type: object
      required:
        - id
        - email
        - username
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: john_doe
        created_at:
          type: string
          format: date-time
          example: '2024-01-15T09:30:00Z'

    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
          description: JWT access token for API requests
        token_type:
          type: string
          enum: [Bearer]
          example: Bearer
        expires_in:
          type: integer
          example: 900
          description: Access token expiration time in seconds
        user:
          $ref: '#/components/schemas/UserResponse'

    RefreshResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
          description: New JWT access token
        token_type:
          type: string
          enum: [Bearer]
          example: Bearer
        expires_in:
          type: integer
          example: 900
          description: Access token expiration time in seconds

    VerifyResponse:
      type: object
      required:
        - valid
        - user_id
        - issued_at
        - expires_at
      properties:
        valid:
          type: boolean
          example: true
          description: Token validity status
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        issued_at:
          type: string
          format: date-time
          example: '2024-01-15T09:30:00Z'
        expires_at:
          type: string
          format: date-time
          example: '2024-01-15T09:45:00Z'
        token_id:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440000
          description: Unique identifier of the token

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error context
            request_id:
              type: string
              format: uuid
              description: Unique request identifier for debugging

  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid request parameters
                  details:
                    field: email
                    reason: Invalid email format
            missing_field:
              value:
                error:
                  code: MISSING_FIELD
                  message: Required field missing
                  details:
                    field: password

    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_credentials:
              value:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
            token_expired:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: Access token has expired
            token_invalid:
              value:
                error:
                  code: TOKEN_INVALID
                  message: Invalid or malformed token
            token_blacklisted:
              value:
                error:
                  code: TOKEN_BLACKLISTED
                  message: Token has been revoked

    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficient_permissions:
              value:
                error:
                  code: INSUFFICIENT_PERMISSIONS
                  message: Admin access required

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            token_not_found:
              value:
                error:
                  code: TOKEN_NOT_FOUND
                  message: Token not found

    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            user_exists:
              value:
                error:
                  code: USER_EXISTS
                  message: User with this email already exists

    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rate_limit:
              value:
                error:
                  code: RATE_LIMIT_EXCEEDED
                  message: Too many requests, please try again later
                  details:
                    retry_after: 60
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              value:
                error:
                  code: INTERNAL_ERROR
                  message: An unexpected error occurred
                  request_id: 550e8400-e29b-41d4-a716-446655440000

  headers:
    X-RateLimit-Limit:
      description: Request limit per minute
      schema:
        type: integer
        example: 5
    X-RateLimit-Remaining:
      description: Remaining requests in current window
      schema:
        type: integer
        example: 4
    X-RateLimit-Reset:
      description: Unix timestamp when rate limit resets
      schema:
        type: integer
        example: 1704456789

tags:
  - name: Authentication
    description: JWT authentication operations